Refactor update code for build process
- Removed `monkey_patching.py`, `setup_logic.py` `setuptools_cxfreeze.py`
- Removed files code merged into `cx_freeze_build.py`
